requirements.txt
requests
beautifulsoup4
ddgs
openai
supabase
python-dotenv
schedule
fastapi
uvicorn
pillow